refactor(kanban): extract board-path/integrity/policy mixins from kanban_db.py (shard s1)
